requirements.txt
pyzmq
Pillow